requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
playwright==1.48.0
pypdf==5.1.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.3.1
pyyaml==6.0.2
tenacity==9.0.0
openpyxl==3.1.5
